App funcionando

+++ Concentracion_Reto/Concentracion_Reto/Context/DB_Context.cs
@@ -47,7 +47,7 @@ namespace Concentracion_Reto.Context
         public int usuarioId { get; set; }
         public string titulo { get; set; }
         public string descripcion { get; set; }
-        public float precio { get; set; }
+        public double precio { get; set; }
         public string ubicacion { get; set; }
     }
