+++ Concentracion_Reto/Concentracion_Reto/Models/dtoPropiedad.cs
@@ -7,7 +7,7 @@ namespace Concentracion_Reto.Models
         public int usuarioId { get; set; }
         public string titulo { get; set; }
         public string descripcion { get; set; }
-        public float precio { get; set; }
+        public double precio { get; set; }
         public string ubicacion { get; set; }
     }
     
@@ -16,7 +16,7 @@ namespace Concentracion_Reto.Models
         public int usuarioId { get; set; }
         public string titulo { get; set; }
         public string descripcion { get; set; }
-        public float precio { get; set; }
+        public double precio { get; set; }
         public string ubicacion { get; set; }
 
     }
